requirements.txt
beautifulsoup4
pyppeteer
fastapi[all]
uvicorn
python-dotenv
boto3
datetime
requests
Pillow
sentencepiece
google-generativeai
transformers
oss2
